examples/applications/cctvs/pyproject.toml
[project]
name = "rohe-cctvs"
version = "0.1.0"
description = "CCTV object detection ML pipeline - ROHE reference application"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "httpx>=0.24.0",
    "Pillow>=9.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rohe-sdk",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]
